updatedPostsの処理をservicesへ切り分ける（feature-akapo_fb20241205）
